Update README.md
Jesus Christ, a central figure in Christianity, was a Jewish preacher and religious leader in the 1st century. Christians believe he is the Son of God and the Messiah, whose life, death, and resurrection offer salvation and eternal life. He is considered the incarnation of God the Son by many Christians.