Update README.md
In Christianity, Jesus is the Son of God as chronicled in the Bible's New Testament, and in most Christian denominations he is held to be God the Son, a prosopon (Person) of the Trinity of God. Christians believe him to be the messiah (giving him the title Christ), who was prophesied in the Bible's Old Testament.

Jesus Christ is the central figure of Christianity, believed by Christians to be the Son of God, the Messiah, and the savior of humanity.  His life, teachings (emphasizing love, forgiveness, and the Kingdom of God), death by crucifixion, and resurrection are foundational to Christian faith.  His impact on history and culture is immense and continues to be studied and debated.